+++ Userland/Libraries/LibAccelGfx/CMakeLists.txt
@@ -3,6 +3,7 @@ include(accelerated_graphics)
 if (HAS_ACCELERATED_GRAPHICS)
     set(SOURCES
         GL.cpp
+        GlyphAtlas.cpp
         Context.cpp
         Painter.cpp
         Program.cpp
